+++ aiobastion/session_management.py
@@ -9,8 +9,14 @@ class SessionManagement:
     # List of attributes from configuration file and serialization
     _SERIALIZED_FIELDS = []
 
-    def __init__(self, epv):
+    def __init__(self, epv, **kwargs):
         self.epv = epv
+
+        _section = "sessionmanagement"
+        _config_source = self.epv.config.config_source
+
+        for _k in kwargs.keys():
+            raise AiobastionConfigurationException(f"Unknown attribute '{_section}/{_k}' in {_config_source}")
 
     async def get_all_connection_components(self):
         """