+++ aiobastion/utilities.py
@@ -30,50 +30,15 @@ class Utilities:
     # List of attributes from configuration file and serialization
     _SERIALIZED_FIELDS = []
 
-    def __init__(self, epv):
+    def __init__(self, epv, **kwargs):
         self.epv = epv
         self.platform = self.Platform(epv)
 
-    @classmethod
-    def _init_validate_class_attributes(cls, serialized: dict, section: str, configfile: str = None) -> dict:
-        """_init_validate_class_attributes      Initialize and validate the Utilities definition (file configuration and serialized)
+        _section = "utilities"
+        _config_source = self.epv.config.config_source
 
-        Arguments:
-            serialized {dict}           Definition from configuration file or serialization
-            section {str}               Verified section name
-
-        Keyword Arguments:
-            configfile {str}            Name of the configuration file
-
-        Raises:
-            AiobastionConfigurationException
-
-        Returns:
-            new_serialized {dict}       Utilities defintion
-        """
-        if not configfile:
-            configfile = "serialized"
-
-        new_serialized = {}
-
-        for k in serialized.keys():
-            keyname = k.lower()
-
-            # # Special validation: integer, boolean
-            # if keyname in ["xxx"]:
-            #     new_serialized[keyname] = validate_integer(configfile, f"{section}/{keyname}", serialized[k])
-
-            if keyname in Utilities._SERIALIZED_FIELDS:
-                # String definition
-                if serialized[k] is not None:
-                    new_serialized[keyname] = serialized[k]
-            else:
-                raise AiobastionConfigurationException(f"Unknown attribute '{section}/{k}' in {configfile}")
-
-        # Default values if not set
-        # new_serialized.setdefault("xxx", Utilities._UTILITIES_DEFAULT_XXX)
-
-        return new_serialized
+        for _k in kwargs.keys():
+            raise AiobastionConfigurationException(f"Unknown attribute '{_section}/{_k}' in {_config_source}")
 
     def to_json(self):
         serialized = {}
@@ -85,7 +50,6 @@ class Utilities:
                 serialized[attr_name] = v
 
         return serialized
-
 
     async def cpm_change_failed_accounts(self, address, username_filter: list = None):
         """