+++ tests/__init__.py
@@ -3,6 +3,7 @@ import os
 
 CONFIG = os.path.join("..", "..", "confs", "config_tests.yml")
 AIM_CONFIG = os.path.join("..", "..", "confs", "config_aim_hp.yml")
+API_USER = "admin_test_restapi"
 
 logging.basicConfig(
     level=logging.DEBUG,