+++ tests/test_accounts.py
@@ -39,9 +39,9 @@ class TestAccount(IsolatedAsyncioTestCase):
         await asyncio.sleep(0)
         # await self.vault.logoff()
 
-    async def get_random_account(self, n=1) -> Union[PrivilegedAccount,List[PrivilegedAccount]]:
+    async def get_random_account(self, n=1, **kwargs) -> Union[PrivilegedAccount,List[PrivilegedAccount]]:
         accounts = await self.vault.account.search_account_by(
-            safe=self.test_safe
+            safe=self.test_safe, **kwargs
         )
         self.assertGreaterEqual(len(accounts), 1)
         if n == 1:
@@ -443,18 +443,29 @@ class TestAccount(IsolatedAsyncioTestCase):
         updated = await self.vault.account.update_single_fc(account, "userName", account.userName)
         self.assertEqual(updated.userName, account.userName)
 
+    async def test_update_platform_account_properties_fc(self):
+        account = await self.get_random_account(platform="WinDesktopLocal")
+
+        updated = await self.vault.account.update_single_fc(account, "Location", "Berlin")
+        self.assertEqual(updated.platformAccountProperties["Location"], "Berlin")
+
+        updated = await self.vault.account.update_single_fc(account, "Location", None)
+        self.assertNotIn("Location",updated.platformAccountProperties)
+
     async def test_update_file_category(self):
-        account = await self.get_random_account()
+        account = await self.get_random_account(platform="WinDesktopLocal", username="wdudhillhd")
         new_username = "tutu"
         new_address = "221.112.152.100"
+        new_location = "Berlin"
+        new_ownername = "Otto Von Bismarck"
         updated = await self.vault.account.update_file_category(account,
-                                                                ["userName", "address"],
-                                                                [new_username, new_address])
+                                                                ["userName", "address", "Location", "OwnerName"],
+                                                                [new_username, new_address, new_location, new_ownername ])
         self.assertEqual(updated.userName, new_username)
         self.assertEqual(updated.address, new_address)
         updated = await self.vault.account.update_file_category(account,
-                                                                ["userName", "address"],
-                                                                [account.userName, account.address])
+                                                                ["userName", "address", "Location", "OwnerName"],
+                                                                [account.userName, account.address, None, None])
         self.assertEqual(updated.userName, account.userName)
         self.assertEqual(updated.address, account.address)